// src/types/cosmeticsType.ts
export interface FortniteType {
  value: string;
  displayValue: string;
  backendValue: string;
}

export interface FortniteRarity {
  value: string;
  displayValue: string;
  backendValue: string;
}

export interface FortniteSeries {
  value: string;
  image: string;
  colors?: string[];
  backendValue?: string;
}

export interface FortniteSet {
  value: string;
  text: string;
}

export interface FortniteImages {
  smallIcon?: string;
  icon?: string;
  featured?: string;
  other?: string | { background?: string };
  lego?: { small?: string; large?: string; wide?: string };
  bean?: { small?: string; large?: string };
  small?: string;
  large?: string;
  wide?: string;
}

export interface FortniteIntroduction {
  chapter?: string;
  season?: string;
  text?: string;
}
// ...
export interface FortniteSingleItem {
  id: string;
  albumArt?: string;
  name: string;
  description?: string;
  type: FortniteType;
  rarity?: FortniteRarity;
  series?: FortniteSeries;
  set?: FortniteSet;
  images: FortniteImages;
  introduction?: FortniteIntroduction;
  added?: string;
}